requirements.txt
Flask==2.2.2 
gunicorn==20.1.0
frozenlist==1.3.3
idna==3.4
Flask-HTTPAuth==4.4.0
Jinja2==3.1.2
pip==23.1.2

mysql-connector-python==8.0.33

idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
pip==23.1.2

requests==2.28.2
